Fix resolution of list type in ExtractAnnotationsTest
Because this test didn't previously include an import for
`java.util.List`, the outer psi type wasn't being properly resolved
(the inner `Integer` type could be resolved since it is from the
`java.lang` package). When the type was being printed, because the
outer type didn't resolve, the source string `List<Integer>` was
used, not using the resolved `java.lang.Integer` type. This caused
a test failure while I was trying to rewrite `TypeItem.toString`.

Test: corrected test
Bug: 309149849
Change-Id: I1bff4ba276182fd49fef774ef2a29c35eb68b286