Fix unhelpful Nullability error message
A method's hasNullabilityInfo is false for a method that is partially
annotated, eg: `@Nullable foo(String param)`.

Because of this, a strange error message was produced when overriding
such a method with "full" nullability info:
@Nullable foo(@NonNull String param), suggesting the child method can
not be annotated Nullable.

Fix this by only considering the return type when validating super
methods in this case, rather than return type + params.

Bug: 309908795
Test: ./gradlew test
Test: android tree checkapi
Change-Id: I033749a89198ad6c87c87bc6a09604aad3f3cada