Merge inner and outer annotation params for toTypeString
I believe the reason these parameters were separate was to prevent
annotations applied to both a type and item from being double
printed. This is no longer an issue for signature files since
annotations are only printed when using kotlin style syntax.

The one place it would still be an issue is stubs. However, just
not printing outer annotations doesn't make sense, as there could
be annotations that are exclusively type-use, not attached to an
item, and these would be left out entirely.

(b/311619997 to track properly printing type-use annotations in stubs)

Bug: 309149849
Test: existing tests pass

Change-Id: I79d7d310db27f36c0634724234bc95747c4441f3