Add test showing type annotations don't appear in stubs
(Or signature files)

This is because the PsiTypePrinter only includes type use annotations
if the constant `SUPPORT_TYPE_USE_ANNOTATIONS` is true, which it isn't.

This wil be updated in followup CLs to allow printing the annotations
in signature files, but to leave the status quo in stubs for now.

Bug: 311619997
Test: Added test

Change-Id: Idb440b2f44d7045f902552c9e168cf85815d9e84