Set type annotations to be included in signatures based on format
Previously, `outerAnnotations=false` and `innerAnnotations=true`. It
seems like the reason for setting `outerAnnotations` to false was to
prevent annotations applied to both the type and item from appearing
twice, but with the new format putting the type in the Kotlin syntax
position, this is no longer a concern.

(Annotations still can't be printed for psi codebases, as mentioned in
prior CLs, to be updated in a following CL)

Bug: 307926200
Test: Migrated existing tests which used type use annotations in signatures.
Updated one to show that type use annotations won't be printed in formats
that don't have the property set. Updated the other to demonstrate parsing
and printing of type use annotations when the property is set.

Change-Id: I29d23e645652e64a8770e140f68e21926fdc6657