Only set super class for stub exception
Previously, when an unknown exception was found in a `throws` clause in
an API signature file it would either load it from another codebase (if
available) or create a stub. In either case it would set the super
class of the returned `ClassItem`, to `java.lang.Throwable`. Even if it
was from another codebase. However, there would be no need to do that
for another codebase as the class's super class would already have been
initialized by that codebase. It could also cause problems changing the
super class of a `ClassItem` in another codebase.

This change only sets the super class when a stub class is created in
the `TextCodebase` being populated.

Bug: 312558482
Test: ./gradlew
Change-Id: I630e08df484f338046132f7e4815ce7d5c13086c