Remove unnecessary code for using  correct super class type
The code was ensuring that the correct parameters were passed to a
generic super class. As the TODO said that information needed to be in
the model and it is now so the code is useless. The code was determined
to be useless by first making it insert the wrong replacement and
ensuring that some tests failed. Then the code was removed completely
and those tests passed. A related TODO was removed as was a TODO about
using raw types in a throws list but that is fine as `Throwable`
sub-classes cannot be generic, enforced by compiler.

Bug: 312558482
Test: ./gradlew
Change-Id: I5f8e6777a11ca4a9695b10f46eea63600cd96b63