Add more tests for StaticFinalBuilder
The `StaticFinalBuilder` check is dependent on whether the `superClass`
is set and so is affected by the following change to stop adding the
first interface type in an interface's `extends` list to `superClass`.
This change adds some more tests to cover those aspects and detect
breakages caused by the following changes.

Bug: 312558482
Test: ./gradlew
Change-Id: I57ecfc3bd5a99b42bd75ee075a551f0249df7aaf