Remove support for a leading @ in annotation patterns
Previously, it was allowed for an annotation pattern passed to options
like `--show-annotation` to have a leading `@`. Unfortunately, that
prevented us from using argument file expansion where `@file` would
cause Clikt to treat the text after `@` as a file name and read its
contents as if it was a set of command line parameters.

Known uses of a leading `@` on an annotation pattern have been cleaned
up to remove the `@` so it should be safe to remove support. If there
are any remaining uses then this and the follow up change will flush
them out and fixing them is easy.

This does not add any tests to make sure that a leading `@` is not
provided as the following up change that enables argument file
expansion will handle that.

Bug: 313398274
Test: ./gradlew
Change-Id: I18bfeb5404386991e7922bfa4f056f27b9c91217