Add support for excluding only some flagged APIs
Most of this change is focused on updating the test to verify that
`--hide-annotation !android.annotation.FlaggedApi("flag")` works as
expected and allows flagged APIs associated with the specific feature
flag to be output even when all other flagged APIs are output.

The other part is to not output the `FlaggedApi` annotation in
signature files if they are hidden. That allows flagged APIs which are
not hidden to be written out without the `FlaggedApi` annotation making
them look just like any other API.

Bug: 313398274
Test: ./gradlew
Change-Id: I06ce82c6f490b90ce6c613e68235985f57bf6c7c