Support excluding annotation patterns
An `AnnotationFilter` consists of a set of annotation patterns. Each
option like `--hide-annotation`, `--show-annotation` that takes an
annotation pattern has a corresponding filter that contains all the
patterns specified using that option.

Previously, annotation patterns could only cause an annotation filter
to match if the pattern matched. There was no way for a specific
pattern that matched to cause the filter to not match, i.e. there
was no way for a pattern to exclude annotations from being matched by
the filter.

This change adds the ability to specify an annotation pattern that
excludes any matched annotations from being matched by the filter even
though there is another pattern in that filter which does match. That
is needed to support hiding all flagged APIs, apart from those for
a specific feature flag.

An excluding annotation pattern is simply one which starts with an `!`.

Bug: 313398274
Test: ./gradlew
Change-Id: Ia1ac94e0caa7c75c739c444d8b95cbae372d7755