Support reverting a flagged item to a previously released item
Previously, `REVERT_UNSTABLE_API` assumed that the item being reverted
was a new item and so revert was equivalent to hiding. This change
extends `REVERT_UNSTABLE_API` to support reverting to a previous item.
It does that by:
1. Adding `revertItem` to `Showability`.
2. Passing that into `ShowOrHide.show(Item)` / `ShowOrHide.hide(Item)`
   methods.
3. Overriding those methods in `REVERT_UNSTABLE_API` to depend on the
   supplied `revertItem`.
4. The `Showability.revertItem` is initialized only when an `Item` has
   a `Showability` of `REVERT_UNSTABLE_API`.

These changes fix the issues with the tests added in the previous
change.

Bug: 314196587
Test: ./gradlew
Change-Id: I5f3111ddf7cba1b20166f41d4ae327370ae69a89