Add test for handling of deprecated in removed files
The propagation of the `Item.deprecated` property is done using an
`ApiVisitor` that does not visit removed APIs. That results in the
deprecated status not being propagated from a removed class onto its
methods. This change adds a test to verify that behavior.

This tests behavior relied upon by the Android build and ensures that
follow up changes do not break it.

Bug: 315206833
Test: ./gradlew
Change-Id: Ib59bd54fa0923fa57b9780776c540b8b1fdbba8e