Propagate effectivelyDeprecated after setting removed
Previously, the `effectivelyDeprecated` property was propagated in the
`visitItem(Item)` method which was called before all the `visit<X>()`
type specific methods. That was a problem as it is necessary to modify
it to take into account the `removed` property. So, this change adds
`markAsDeprecatedIfNonPackageParentIsDeprecated` and calls it
explicitly at the end of the `visit<X>()` methods in preparation for a
follow up change that will factor in the `removed` value.

Bug: 315206833
Test: ./gradlew
Change-Id: I2bdd0622adb5c7d1199531db4c7b64d799e2e61e