Rename HIDE_UNSTABLE_API to REVERT_UNSTABLE_API
This more closely matches the required semantics for items annotated
with the `@FlaggedApi` when they are not to be included in the
generated API.

A new item that is annotated with `@FlaggedApi` needs to be hidden from
the API, but an existing item that is changed needs to be reverted to
its previous form.

Follow up changes will add support for that new behavior.

Bug: 314196587
Test: ./gradlew
Change-Id: I7597dbecf50c5fda636204aaca9aa3f91711f4d1