Add TraversingVisitor for more efficient processing
The `BaseItemVisitor` tries to be a one size fits all base class for
visitors which makes it more complex than some uses and does not
provide the capabilities needed by others. This change adds the
`TraversingVisitor` for use by those `BaseItemVisitor` implementations
which need to simply do the same thing for each `Item` and which need
to have some control over the traversal that is being performed.

Bug: 316212602
Test: ./gradlew
Change-Id: Idc5d7610571cfeb83e878476420e264d4059cb91