Only report UnmatchedMergeAnnotation when there are annotations to copy
Previously, it would report `UnmatchedMergeAnnotation` for every item
in the external `Codebase` that did not have a corresponding item in
the main `Codebase`. That can lead to errors being reported that may be
tedious to fix. This change simply ignores unmatched items if there are
no annotations to copy and sets the default severity of
`UnmatchedMergeAnnotation` to error.

That required adding some additional annotations to a couple of tests
and specifying "--warning UnmatchedMergeAnnotation` to preserve their
behavior.

Bug: 316212602
Test: ./gradlew
Change-Id: I0c98ab1fe20ee3374ffca8b9a37d6cc7350090dd