Add default toString implementation for TypeItem
This changes the behavior for TextTypeItem, but only when the parsed
type contains annotations or kotlin nulls. There are currently no
real use cases of annotations on type strings, and the final kotlin
null suffix of a type isn't included in the `type` string previously
used for `TextTypeItem.toString` (that's changed in the a later CL).

Bug: 309149849
Test: existing tests pass.
Change-Id: I6faae066eb09f34304851145c4212eba53de284c