Simplify the BaselineCheck construction and use
Previously, the `BaselineCheck` was passed the `BaselineInfo` and then
in its `apply()` method had to make a choice as to which of the String
properties in there it would use. There was similar logic in
`buildBaselineCheck` to do something similar for the inputs.

This change combines the logic into the `buildBaselineCheck` so it
chooses which options to provide and which contents to use and stores
that in the `BaselineCheck` and then `apply()` just has to compare the
actual file contents with the expected file contents.

Bug: 316456211
Test: ./gradlew
Change-Id: I0a9ad41abc1727290e9639b527998716b4c7ec44