Filter KotlinInteropChecks like ApiLint checks
Previously, `KotlinInteropChecks` would report issues directly to the
`Reporter`. That differed from `ApiLint` which would pass its issues
through its own `ApiLint.report(...)` method which would filter the
issues in a couple of ways to make sure that they are relevant to the
API surface being generated.

This change extracts the `ApiLint.report(...)` method into a
`FilteringReporter` class which it can pass to `KotlinInteropChecks` so
it can filter issues in the same way.

Bug: 315327830
Bug: 312744792
Bug: 311690008
Test: ./gradlew
Change-Id: Ib02ba2668be8806a64fea2570912e6499a7408a8